Add logback for logging to test Listener. Create logback.xml configuration. Rename user tests classes.
