--- UserTests.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="User Tests">
-    <test name="User Tests">
-        <classes>
-            <class name="api.user.UserClientPositiveTest"/>
-            <class name="api.user.UserClientNegativeTest"/>
-        </classes>
-    </test>
-</suite>
