--- src/logs/package-info.java
@@ -1,3 +1,0 @@
-/**
- * This package containes log report files to each test method
- */
