+++ src/main/java/client/BaseClient.java
@@ -3,10 +3,7 @@ package client;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.config.LogConfig;
 import io.restassured.config.RestAssuredConfig;
-import io.restassured.filter.log.ErrorLoggingFilter;
 import io.restassured.filter.log.LogDetail;
-import io.restassured.filter.log.RequestLoggingFilter;
-import io.restassured.filter.log.ResponseLoggingFilter;
 import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
 
@@ -14,20 +11,19 @@ public abstract class BaseClient {
     private final String baseUrl = "https://petstore.swagger.io/v2";
 
     protected final RequestSpecification baseRequestSpecification(final ContentType contentType, LogDetail logDetail) {
-        
         RequestSpecBuilder builder = new RequestSpecBuilder()
                 .setConfig(RestAssuredConfig.config().logConfig(LogConfig.logConfig().enableLoggingOfRequestAndResponseIfValidationFails()))
                 .setConfig(RestAssuredConfig.config().logConfig(LogConfig.logConfig().enablePrettyPrinting(true)))
                 .setBaseUri(baseUrl)
                 .setAccept(contentType)
                 .setContentType(contentType);
-        if(logDetail != null) {
+
+        if (logDetail != null) {
             builder.log(logDetail);
         }
         return builder.build();
     }
 
-    //additional version of method with default LogDetail : 'ALL', ContentType : 'JSON'
     protected final RequestSpecification baseRequestSpecification(final ContentType contentType) {
         return baseRequestSpecification(contentType, null);
     }
