+++ src/test/java/api/user/CreateUserTests.java
@@ -5,18 +5,28 @@ import assertion.UserAssertion;
 import builders.UserBuilder;
 import client.UserClient;
 import io.restassured.response.Response;
+import listeners.Listener;
 import model.User;
 import org.testng.Assert;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.DataProvider;
+import org.testng.annotations.Listeners;
 import org.testng.annotations.Test;
 
 import java.net.HttpURLConnection;
+import java.util.ArrayList;
 import java.util.List;
 
-public class UserClientCreateTest {
+@Listeners(Listener.class)
+public class CreateUserTests {
     UserClient userClient;
     UserBuilder userBuilder;
+
+    @BeforeClass
+    public void beforeClass() {
+        userClient = new UserClient();
+        userBuilder = new UserBuilder();
+    }
 
     @Test
     public void createTest() {
@@ -25,15 +35,25 @@ public class UserClientCreateTest {
         BaseAssertion.checkResponse(response, HttpURLConnection.HTTP_OK);
     }
 
-    @BeforeClass
-    public void beforeClass() {
-        userClient = new UserClient();
-        userBuilder = new UserBuilder();
+    @Test
+    public void createWithList() {
+        final List<User> randomUsers = userBuilder.constructRandomListValidUsers(3);
+        final Response response = userClient.createWithList(randomUsers);
+
+        BaseAssertion.checkResponse(response, HttpURLConnection.HTTP_OK);
+    }
+
+    @Test
+    public void createByEmptyList() {
+        Response response = userClient.createWithList(new ArrayList<>());
+
+        BaseAssertion.checkResponse(response, HttpURLConnection.HTTP_OK);
     }
 
     @Test(dataProvider = "users")
-    public void test(User user) {
+    public void createRandomUser(User user) {
         Response response = userClient.create(user);
+
         Assert.assertEquals(response.jsonPath().getLong("message"), user.getId());
     }
 
@@ -51,13 +71,13 @@ public class UserClientCreateTest {
         User returnUser = response.as(User.class);
         UserAssertion.checkValidUser(user, returnUser);
     }
+
     @DataProvider
     public Object[] getCreatingUsers() {
         List<User> users = userBuilder.constructRandomListValidUsers(4);
-        for(User u : users) {
+        for (User u : users) {
             userClient.create(u);
         }
-        return  users.toArray();
+        return users.toArray();
     }
-
 }
